requirements.txt
PyQt5==5.14.1
RPi.GPIO==0.7.0
configparser==4.0.2